ade code: quality synthesis — digit-picker fixes, app.tsx extractions, UI polish
