+++ apps/ade-cli/src/tuiClient/__tests__/remoteLauncher.test.ts
@@ -1,22 +1,12 @@
-import { EventEmitter } from "node:events";
-import { PassThrough } from "node:stream";
 import { describe, expect, it, vi } from "vitest";
 import {
   buildRemoteRuntimeRpcCommand,
   listRemoteSessions,
   parseRemoteAdeCodeArgs,
-  ProcessJsonRpcClient,
   remoteRuntimeLayoutCandidates,
   selectProject,
   takeAdeCodeRemoteArgs,
 } from "../remoteLauncher";
-
-class FakeRemoteRpcChild extends EventEmitter {
-  readonly stdout = new PassThrough();
-  readonly stderr = new PassThrough();
-  readonly stdin = new PassThrough();
-  readonly kill = vi.fn();
-}
 
 describe("ade code remote launcher", () => {
   it("detects remote as the first standalone ade code positional", () => {
@@ -62,28 +52,6 @@ describe("ade code remote launcher", () => {
     expect(buildRemoteRuntimeRpcCommand(layout)).toContain("export ADE_DISABLE_RUNTIME_SERVICE_INSTALL=1");
     expect(buildRemoteRuntimeRpcCommand(layout)).toContain('export ADE_PTY_HOST_WORKER_COMMAND="$HOME/.ade-beta/bin/ade"');
     expect(buildRemoteRuntimeRpcCommand(layout)).toContain("exec $HOME/.ade-beta/bin/ade --socket $HOME/.ade-beta/sock/ade.sock rpc --stdio");
-  });
-
-  it("clears pending RPC timers when remote responses resolve or reject", async () => {
-    const clearTimeoutSpy = vi.spyOn(global, "clearTimeout");
-    const child = new FakeRemoteRpcChild();
-    const client = new ProcessJsonRpcClient(child as never);
-
-    const resolved = client.request("ade/ping");
-    child.stdout.write(`${JSON.stringify({ jsonrpc: "2.0", id: 1, result: { ok: true } })}\n`);
-    await expect(resolved).resolves.toEqual({ ok: true });
-
-    const rejected = client.request("ade/fail");
-    child.stdout.write(`${JSON.stringify({
-      jsonrpc: "2.0",
-      id: 2,
-      error: { code: -32000, message: "remote failed" },
-    })}\n`);
-    await expect(rejected).rejects.toThrow("remote failed");
-
-    expect(clearTimeoutSpy).toHaveBeenCalledTimes(2);
-    client.close();
-    clearTimeoutSpy.mockRestore();
   });
 
   it("attaches to stable remote runtimes without repairing or installing services", () => {
